+++ instrumentation/net/http/otelhttp/common.go
@@ -12,9 +12,16 @@ import (
 
 // Attribute keys that can be added to a span.
 const (
-	ReadBytesKey  = attribute.Key("http.read_bytes")  // if anything was read from the request body, the total number of bytes read
-	ReadErrorKey  = attribute.Key("http.read_error")  // If an error occurred while reading a request, the string of the error (io.EOF is not recorded)
+	// Deprecated: use semconv.HTTPRequestBodySizeKey instead.
+	ReadBytesKey = attribute.Key("http.read_bytes") // if anything was read from the request body, the total number of bytes read
+
+	// Deprecated: use semconv.ErrorMessageKey instead.
+	ReadErrorKey = attribute.Key("http.read_error") // If an error occurred while reading a request, the string of the error (io.EOF is not recorded)
+
+	// Deprecated: use semconv.HTTPResponseBodySizeKey instead.
 	WroteBytesKey = attribute.Key("http.wrote_bytes") // if anything was written to the response writer, the total number of bytes written
+
+	// Deprecated: use semconv.ErrorMessageKey instead.
 	WriteErrorKey = attribute.Key("http.write_error") // if an error occurred while writing a reply, the string of the error (io.EOF is not recorded)
 )